+++ EVEMon.Common/MarketOrder.cs
@@ -18,7 +18,7 @@ namespace EVEMon.Common
 
         private OrderState m_state;
 
-        private readonly long m_itemID;
+        private readonly int m_itemID;
 
 
         #region Constructors
@@ -295,10 +295,10 @@ namespace EVEMon.Common
         /// </summary>
         /// <param name="src"></param>
         /// <returns></returns>
-        private static long GetItemID(SerializableOrderBase src)
+        private static int GetItemID(SerializableOrderBase src)
         {
             // Try get item ID by source
-            long itemID = src.ItemID;
+            int itemID = src.ItemID;
 
             // We failed? Try get item ID by name
             if (itemID == 0)
@@ -327,7 +327,7 @@ namespace EVEMon.Common
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        private static Station GetStationByID(long id)
+        private static Station GetStationByID(int id)
         {
             // Look for the station in datafile, if we fail then it may be a conquerable outpost station
             Station station = StaticGeography.GetStationByID(id) ?? ConquerableStation.GetStationByID(id);
