+++ Tools/XmlGenerator/StaticData/InvMarketGroup.cs
@@ -5,16 +5,16 @@ namespace EVEMon.XmlGenerator.StaticData
     public sealed class InvMarketGroup : IHasID
     {
         [XmlElement("marketGroupID")]
-        public long ID { get; set; }
+        public int ID { get; set; }
 
         [XmlElement("marketGroupName")]
         public string Name { get; set; }
 
         [XmlElement("parentGroupID")]
-        public long? ParentID { get; set; }
+        public int? ParentID { get; set; }
 
         [XmlElement("iconID")]
-        public long? IconID { get; set; }
+        public int? IconID { get; set; }
 
         [XmlElement("description")]
         public string Description { get; set; }
